SMGCPA: Accelerate constructing of a new SMG by copying same hasValueEdges at SMGJoinSubSMGs. Simplify getNullEdgesMapOffsetToSizeForObject() because SMGHasValueEdges already merges neighboring zero value fields
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/SMGCPA-merge-hvEdges@32825 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/graphs/SMGTest.java
@@ -90,8 +90,8 @@ public class SMGTest {
     SMGEdgeHasValue hv = new SMGEdgeHasValue(mockTypeSize, 32, obj1, SMGZeroValue.INSTANCE);
     smg1.addHasValueEdge(hv);
 
-    TreeMap<Long, Integer> nullEdges = smg1.getNullEdgesMapOffsetToSizeForObject(obj1);
-    assertThat(nullEdges).containsExactly(32L, 32);
+    TreeMap<Long, Long> nullEdges = smg1.getNullEdgesMapOffsetToSizeForObject(obj1);
+    assertThat(nullEdges).containsExactly(32L, 32L);
   }
 
   @Test
