+++ src/org/sosy_lab/cpachecker/cpa/smg/graphs/UnmodifiableSMG.java
@@ -76,7 +76,7 @@ public interface UnmodifiableSMG {
 
   int getSizeofPtrInBits();
 
-  TreeMap<Long, Integer> getNullEdgesMapOffsetToSizeForObject(SMGObject pObj);
+  TreeMap<Long, Long> getNullEdgesMapOffsetToSizeForObject(SMGObject pObj);
 
   boolean isPointer(SMGValue value);
